facui double

+++ QuickSort/quicksort.cpp
@@ -1,7 +1,7 @@
 #include <vector>
 #include <fstream>
 
-std::vector<int> quicksort(std::vector<int>& arr)
+std::vector<double> quicksort(std::vector<double>& arr)
 {
     if (arr.size() <= 1)
     {
@@ -9,9 +9,9 @@ std::vector<int> quicksort(std::vector<int>& arr)
     }
     else
     {
-        int pivot = arr[arr.size()/3];  ///am ales un pivot random, o sa schimbam asta sau facem mai multe teste cu asta
-        std::vector<int> left;
-        std::vector<int> right;
+        double pivot = arr[arr.size()/3];  ///am ales un pivot random, o sa schimbam asta sau facem mai multe teste cu asta
+        std::vector<double> left;
+        std::vector<double> right;
         for (int i = 1; i < arr.size(); i++)    ///impart vectorul initial in doua astfel incat elementele din left sa fie mai mici decat pivotul si din right mai mare decat pivotul
         {
             if (arr[i] < pivot)
@@ -41,7 +41,7 @@ int main(int argc, char *argv[])
     std::ifstream fin(argv[1]);
     std::ofstream fout(argv[2]);
 
-    std::vector<int> v;
+    std::vector<double> v;
     int num;
     int temp;
     fin >> temp;
